add new controllers

+++ InfoTestMe.Admin.Web/Controllers/CoursesController.cs
@@ -5,6 +5,7 @@ using InfoTestMe.Common.Models;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
@@ -19,11 +20,11 @@ namespace InfoTestMe.Admin.Web.Controllers
         private readonly IUserService _userService;
         private readonly IAuthorService _authorService;
         private readonly ICourseService _courseService;
-        private readonly ICourseThemeService _courseThemeService;
         private readonly ICoursePageService _coursePageService;
 
         public CoursesController(InfoTestMeDataContext db)
         {
+            _userService = new UserService(db);
             _courseService = new CourseService(db);
             _coursePageService = new CoursePageService(db);
             _authorService = new AuthorService(db);
@@ -78,127 +79,6 @@ namespace InfoTestMe.Admin.Web.Controllers
             return BadRequest();
         }
 
-        /*/// <summary>
-        /// Получение тем курса
-        /// </summary>
-        /// <returns></returns>
-        [HttpGet("{courseId}")]
-        public async Task<ActionResult<IEnumerable<CourseThemeDTO>>> GetCourseThemes(int courseId)
-        {
-            if (courseId != 0)
-            {
-                var result = await _courseThemeService.GetAllThemeByCourseId(courseId);
-
-                return result == null ? NotFound() : Ok(result);
-            }
-            return BadRequest();
-        }*/
-
-        /// <summary>
-        /// Получение страниц курса по id темы
-        /// </summary>
-        /// <param name="id"></param>
-        /// <returns></returns>
-        [HttpGet("theme/{themeId}")]
-        public async Task<ActionResult<IEnumerable<CoursePageDTO>>> GetAllCoursePagesByTheme(int themeId)
-        {
-            if (themeId != 0)
-            {
-                var result = await _coursePageService.GetPagesByThemeId(themeId);
-
-                return result == null ? NotFound() : Ok(result);
-            }
-            return BadRequest();
-        }
-
-        /// <summary>
-        /// Создание темы
-        /// </summary>
-        /// <param name="id"></param>
-        /// <returns></returns>
-        [HttpPost("theme")]
-        public ActionResult<int> CreateTheme(CourseThemeDTO themeDTO)
-        {
-            if (themeDTO != null)
-            {
-                int id = _courseThemeService.CreateThemeAndGetId(themeDTO);
-
-                return Ok(id);
-            }
-            return BadRequest();
-        }
-
-        /// <summary>
-        /// Получение страницы курса по id страницы
-        /// </summary>
-        /// <param name="id"></param>
-        /// <returns></returns>
-        [HttpGet("page/{id}")]
-        public ActionResult<CoursePageDTO> GetCoursePage(int id)
-        {
-            if (id != 0)
-            {
-                var result = _coursePageService.Get(id);
-
-                return result == null ? NotFound() : Ok(result);
-            }
-            return BadRequest();
-        }
-
-        /// <summary>
-        /// Создание страницы курса
-        /// </summary>
-        /// <param name="id"></param>
-        /// <returns></returns>
-        [Authorize(Roles = "Author")]
-        [HttpPost("page")]
-        public IActionResult CreateCoursePage([FromBody] CoursePageDTO pageDTO)
-        {
-            if (pageDTO != null)
-            {
-                var result = _coursePageService.Create(pageDTO);
-
-                return result ? Ok() : StatusCode(500);
-            }
-            return BadRequest();
-        }
-
-        /// <summary>
-        /// Обновление страницы курса
-        /// </summary>
-        /// <param name="id"></param>
-        /// <returns></returns>
-        [Authorize(Roles = "Author")]
-        [HttpPatch("page")]
-        public IActionResult UpdateCoursePage([FromBody] CoursePageDTO pageDTO)
-        {
-            if (pageDTO != null)
-            {
-                var result = _coursePageService.Update(pageDTO);
-
-                return result ? Ok() : StatusCode(500);
-            }
-            return BadRequest();
-        }
-
-        /// <summary>
-        /// Удаление страницы курса
-        /// </summary>
-        /// <param name="id"></param>
-        /// <returns></returns>
-        [Authorize(Roles = "Author")]
-        [HttpDelete("page/{pageId}")]
-        public IActionResult DeleteCoursePage(int pageId)
-        {
-            if (pageId != 0)
-            {
-                var result = _coursePageService.Delete(pageId);
-
-                return result ? Ok() : StatusCode(500);
-            }
-            return BadRequest();
-        }
-
         /// <summary>
         /// Создание курса
         /// </summary>
@@ -232,11 +112,11 @@ namespace InfoTestMe.Admin.Web.Controllers
             { 
                 //get current author
                 Author author = _authorService.GetAuthorByLogin(HttpContext.User.Identity.Name);
-                courseDTO.AuthorId = author.Id;
-
-                var result = _courseService.Update(courseDTO);
-
-                return result ? Ok() : StatusCode(500);
+                if(courseDTO.AuthorId == author.Id)
+                {
+                    var result = _courseService.Update(courseDTO);
+                    return result ? Ok() : StatusCode(500);
+                }
             }
             return BadRequest();
         }
@@ -251,9 +131,13 @@ namespace InfoTestMe.Admin.Web.Controllers
         {
             if (id != 0)
             {
-                var result = _courseService.Delete(id);
-
-                return result ? Ok() : StatusCode(500);
+                //get current author
+                AuthorDTO author = _authorService.GetAuthorDTOByLogin(HttpContext.User.Identity.Name);
+                if(author.Courses.Select(c => c.Id).Contains(id))
+                {
+                    var result = _courseService.Delete(id);
+                    return result ? Ok() : StatusCode(500);
+                }
             }
             return BadRequest();
         }
